--- pyg.py
@@ -1,1 +1,0 @@
-#     return x, mapping
